Revise definition of lcs, improve error reporting
The lcs of "enhanced" string type and a corresponding primitive
type is that primitive type.

Rationale: the annotated string type (containing a value that can be
parsed as a number) can only be the type of that particular string.

From this follows that combination of such string type and a primitive
type can be either a string type (which denies the purpose of
enhanced string type), or a primitive type.